React cleanup + error handling + notification messages
